Improve solidity loop style

+++ contracts/types/IngotSpec.sol
@@ -30,7 +30,7 @@ library IngotSpecLib {
 
     function getName(IngotSpec memory _ingotSpec) public view returns (string memory) {
         string memory symbol = "Ingot";
-        for (uint i = 0; i < _ingotSpec.nuggetSpecs.length; ++i) {
+        for (uint256 i = 0; i < _ingotSpec.nuggetSpecs.length; ++i) {
             NuggetSpec memory _nuggetSpec = _ingotSpec.nuggetSpecs[i];
             symbol = string.concat(symbol, " ", _nuggetSpec.getNameSuffix());
         }
@@ -39,7 +39,7 @@ library IngotSpecLib {
 
     function getSymbol(IngotSpec memory _ingotSpec) public view returns (string memory) {
         string memory symbol = "IO";
-        for (uint i = 0; i < _ingotSpec.nuggetSpecs.length; ++i) {
+        for (uint256 i = 0; i < _ingotSpec.nuggetSpecs.length; ++i) {
             NuggetSpec memory _nuggetSpec = _ingotSpec.nuggetSpecs[i];
             symbol = string.concat(symbol, " ", _nuggetSpec.getSymbolSuffix());
         }
